#23: Complete HowTo unit test
HowTo tests were failing because v-playback had not been migrated, and
howto.js was not managing videos correctly. That was fixed in a previous
commit, this fleshes out the HowTo unit tests to account for the fixed
functionality. The test has errors from video.js, which is used by
v-playback. It seems like the testing harness is not providing valid
DOM elements for display. Nonetheless, the tests pass, and when viewed,
the page appears and loads videos just fine. I attempted to find the
source of these errors, but wasn't able to.