#66: Refactored getArticleMetadata for easier mock
Mocking getArticleMetadata easily required moving it into an enclosing
object so that the manual mock could do the same, allowing it to access
properties. The manual mock was written following along with
[jest's manual mock](https://jestjs.io/docs/manual-mocks).

The mock of the resulting getArticle also provides a stub of the
SemanticScholar API response as of 2/2/22, for reference and testing.