#132: Add MCR specific tool page
Renamed tool template to specify that it is MCR tool specific -- at
least given the tools we've added so far. Added a dependency for
EditImage.vue, a component of ImageUpload.vue that is itself a component
of MCRToolTemplate.vue. Removed the old MCR homepage.

Added a pageContent getter for tools as well.