#25: Initial SmilesCanvas unit test
For some reason, SmilesCanvas is not able to reach the same document as
the jest harness is providing the test, so it errors out when trying to
render the SMILES output. Haven't been able to debug, this is a
placeholder test that I'll come back to when I'm able to figure it out.