#132 write slot driven templates (#142)
* #132: Add MCR specific tool page

Renamed tool template to specify that it is MCR tool specific -- at
least given the tools we've added so far. Added a dependency for
EditImage.vue, a component of ImageUpload.vue that is itself a component
of MCRToolTemplate.vue. Removed the old MCR homepage.

Added a pageContent getter for tools as well.

* #132: tag auth mock more obviously

* #132: Use MCRToolTemplate in all tools, add socket

Job Manager uses vue-socket.io -- copied config from old repo and added
this and socket.io-client as well. Updated MCRToolTemplate to account
for move from state to components -- added card, moved reference
container.

* #132: Remove old ToolTemplate

* #132: Update MCRToolTemplate to use tool prop

Also add spinner for loading results

* #132: Fix typos

* #132: Add optional mocks to test wrapper

`$socket` needed to be mocked for some tests, added that functionality
to `createWrapper`.

* #132: Write Tool Card that uses slots, not props

* #132: Move major tool page content to slots

Not all content could be moved, as some is needed for job management,
but major display content was turned into slots for template HTML.

* #132: Move Toolset content from data to slots

* #132: Wrote tests of new templates

Also updated the wrapper function to allow for user-defined stubs

* #132: Remove replacement-tool-page slot

Was created for Dynamfit, ChemProps, but those really should be separate
components, so I removed the slot.

* #132: Make tool cards full-width

* #132: Convert Module Tools page to new system

* #133: Convert Binarization tools and homepage

Moved tool data from the `data` field to slots in the tool page template
and moved the homepage data from `data` to the toolset page template
slots.

* #133: Fix seeming typo in ImageUpload.vue

I may be wrong, but the way the selects prop is used in the template
seems like it should be an Array, not an Object. All the selects fields
in each tool are Arrays as far as I remember. Certainly for the
binarization tools.

* #134: Convert Characterization tools

* #135: Convert reconstruction tools

* #136: Convert Intelligent characterization tool

* #139: Remove ChemProps, Dynamfit from template

ChemProps and Dynamfit rewrite all the job management code already, so
they should really be separate components.

* #139: Wrap lifecycle events in `this.card` checks

If a component is being rendered as a card, it doesn't need to check for
authorization or commit its header. Also removed unused minWidth param
from renderDialog.

* #132: Convert MCR homepage

* #132: Remove unnecessary index files

ToolSetTemplate no longer needs to keep all tools as components, so the
index files that centralized them were removed.

* #132: Tweak ToolCard size, layout

* #132: Write test for MCRToolTemplate

* #132: Remove typos and testing code

* #132: Attempt at cleaning up scoped styles

* #132: Migrate SimulationTools, Polymerizer

Add fallback content for toolset pages, if description is not present.

* #132: Change submit button implementation

You can't listen to events on slots, so the button was moved to the
tool template, and the slot only provides the button text now.

* #132: Typo in styles import

* #132: Fix MathML syntax

* #132: Move tool, toolset routes to own module

* #132: Replace MathML with MathJax

Includes a CDN script tag in index.html, and a new component, but now
math will render in all browsers.

* #132: Combine tool scss files into one

* #132: Fix style imports for new consolidated file

* #132: Updated vue-advanced-cropper version

* #132: Wrap styles to avoid overflow